Move logic into a cheater
