+++ protostar/testing/cairo_cheatcodes/declare_cairo_cheatcode.py
@@ -1,24 +1,9 @@
 import asyncio
-from dataclasses import dataclass
-from pathlib import Path
 from typing import Any, Protocol
-
-from starkware.python.utils import from_bytes
-from starkware.starknet.business_logic.transaction.objects import InternalDeclare
-from starkware.starknet.public.abi import AbiType
-from starkware.starknet.services.api.gateway.transaction import (
-    DEFAULT_DECLARE_SENDER_ADDRESS,
-)
-from starkware.starknet.testing.contract import DeclaredClass
-from starkware.starknet.testing.contract_utils import EventManager, get_abi
 
 from protostar.compiler import ProjectCompiler
 from protostar.testing.cairo_cheatcodes.cairo_cheatcode import CairoCheatcode
-
-
-@dataclass
-class DeclaredContract:
-    class_hash: int
+from protostar.contract_types import DeclaredContract
 
 
 class DeclareCheatcodeProtocol(Protocol):
@@ -43,60 +28,19 @@ class DeclareCairoCheatcode(CairoCheatcode):
         return self.declare
 
     def declare(self, contract: str) -> DeclaredContract:
-        declared_class = asyncio.run(self._declare_contract(contract))
+        compiled_contract = (
+            self._project_compiler.compile_contract_from_contract_identifier(contract)
+        )
+        declared_class = asyncio.run(
+            self.cheaters.contracts.declare_contract(compiled_contract)
+        )
 
         assert declared_class
         class_hash = declared_class.class_hash
 
-        self.cheatable_state.class_hash_to_contract_abi_map[
-            class_hash
-        ] = declared_class.abi
-
-        self.cheatable_state.class_hash_to_contract_path_map[class_hash] = Path(
-            contract
+        self.cheaters.contracts.bind_class_hash_to_contract_identifier(
+            class_hash=class_hash,
+            contract_identifier=contract,
         )
 
         return DeclaredContract(class_hash)
-
-    async def _declare_contract(self, contract: str):
-        contract_class = (
-            self._project_compiler.compile_contract_from_contract_identifier(contract)
-        )
-
-        starknet_config = self.starknet.state.general_config
-
-        tx = InternalDeclare.create(
-            contract_class=contract_class,
-            chain_id=starknet_config.chain_id.value,
-            sender_address=DEFAULT_DECLARE_SENDER_ADDRESS,
-            max_fee=0,
-            version=0,
-            signature=[],
-            nonce=0,
-        )
-
-        with self.cheatable_state.copy_and_apply() as state_copy:
-            await tx.apply_state_updates(
-                state=state_copy, general_config=starknet_config
-            )
-
-        abi = get_abi(contract_class=contract_class)
-        self._add_event_abi_to_state(abi)
-        class_hash = tx.class_hash
-        assert class_hash is not None
-        await self.cheatable_state.set_contract_class(class_hash, contract_class)
-
-        return DeclaredClass(
-            class_hash=from_bytes(class_hash),
-            abi=get_abi(contract_class=contract_class),
-        )
-
-    def _add_event_abi_to_state(self, abi: AbiType):
-        event_manager = EventManager(abi=abi)
-        self.cheatable_state.update_event_selector_to_name_map(
-            # pylint: disable=protected-access
-            event_manager._selector_to_name
-        )
-        # pylint: disable=protected-access
-        for event_name in event_manager._selector_to_name.values():
-            self.cheatable_state.event_name_to_contract_abi_map[event_name] = abi
