+++ protostar/testing/cheatcodes/declare_cheatcode.py
@@ -1,5 +1,4 @@
 import asyncio
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Protocol
 
@@ -13,12 +12,8 @@ from starkware.starknet.testing.contract import DeclaredClass
 from starkware.starknet.testing.contract_utils import EventManager, get_abi
 
 from protostar.compiler import ProjectCompiler
+from protostar.contract_types import DeclaredContract
 from protostar.starknet import Cheatcode, KeywordOnlyArgumentCheatcodeException
-
-
-@dataclass
-class DeclaredContract:
-    class_hash: int
 
 
 class DeclareCheatcodeProtocol(Protocol):
